+++ verik-importer/src/test/kotlin/io/verik/importer/cast/cast/DescriptorCasterTest.kt
@@ -74,7 +74,7 @@ internal class DescriptorCasterTest : BaseTest() {
             """
                 logic [1:0] x;
             """.trimIndent(),
-            "Property(x, BitDescriptor(Nothing, LiteralExpression(1), LiteralExpression(0), 0))"
+            "Property(x, BitDescriptor(Nothing, LiteralExpression(*), LiteralExpression(*), 0))"
         ) { it.findDeclaration("x") }
     }
 
@@ -85,7 +85,7 @@ internal class DescriptorCasterTest : BaseTest() {
             """
                 logic [1:0][1:0] x;
             """.trimIndent(),
-            "Property(x, PackedDescriptor(Nothing, BitDescriptor(*), LiteralExpression(1), LiteralExpression(0)))"
+            "Property(x, PackedDescriptor(Nothing, BitDescriptor(*), LiteralExpression(*), LiteralExpression(*)))"
         ) { it.findDeclaration("x") }
     }