+++ verik-importer/src/test/kotlin/io/verik/importer/cast/cast/ExpressionCasterTest.kt
@@ -30,7 +30,7 @@ internal class ExpressionCasterTest : BaseTest() {
             """
                 logic [N:0] x;
             """.trimIndent(),
-            "Property(x, BitDescriptor(Nothing, ReferenceExpression(N, Nothing), LiteralExpression(0), 0))"
+            "Property(x, BitDescriptor(Nothing, ReferenceExpression(Nothing, N, Nothing), LiteralExpression(*), 0))"
         ) { it.findDeclaration("x") }
     }
 
@@ -44,7 +44,7 @@ internal class ExpressionCasterTest : BaseTest() {
             """
                 Property(
                     x, ReferenceDescriptor(Nothing, c, Nothing,
-                    [ExpressionTypeArgument(null, ReferenceExpression(d, Nothing))])
+                    [ExpressionTypeArgument(null, ReferenceExpression(Nothing, d, Nothing))])
                 )
             """.trimIndent()
         ) { it.findDeclaration("x") }